Encode(ROSpan<byte>) and Decode(ROSpan<char>) moved from base class to children => allows inlineing
