+++ source/gfoidl.Base64/Internal/Base64UrlEncoder/Base64UrlEncoder.Decode.cs
@@ -30,6 +30,25 @@ namespace gfoidl.Base64.Internal
         }
         //---------------------------------------------------------------------
         // PERF: can't be in base class due to inlining (generic virtual)
+        public override byte[] Decode(ReadOnlySpan<char> encoded)
+        {
+            if (encoded.IsEmpty) return Array.Empty<byte>();
+
+            int dataLength         = this.GetDecodedLength(encoded);
+            byte[] data            = new byte[dataLength];
+            OperationStatus status = this.DecodeImpl(encoded, data, out int consumed, out int written, dataLength);
+
+            if (status == OperationStatus.InvalidData)
+                ThrowHelper.ThrowForOperationNotDone(status);
+
+            Debug.Assert(status         == OperationStatus.Done);
+            Debug.Assert(encoded.Length == consumed);
+            Debug.Assert(data.Length    == written);
+
+            return data;
+        }
+        //---------------------------------------------------------------------
+        // PERF: can't be in base class due to inlining (generic virtual)
         protected override OperationStatus DecodeCore(
             ReadOnlySpan<byte> encoded,
             Span<byte> data,
@@ -49,6 +68,7 @@ namespace gfoidl.Base64.Internal
             bool isFinalBlock = true)
             => this.DecodeImpl(encoded, data, out consumed, out written, decodedLength, isFinalBlock);
         //---------------------------------------------------------------------
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private OperationStatus DecodeImpl<T>(
             ReadOnlySpan<T> encoded,
             Span<byte> data,
